requirements.txt
openai==2.15.0
python-dotenv==1.2.1
gradio==6.3.0
pytest
